+++ src/lzApp/libs/ExcessivelySafeCall.sol
@@ -2,7 +2,7 @@
 pragma solidity >=0.7.6;
 
 library ExcessivelySafeCall {
-    uint constant LOW_28_MASK = 0x00000000ffffffffffffffffffffffffffffffffffffffffffffffffffffffff;
+    uint256 constant LOW_28_MASK = 0x00000000ffffffffffffffffffffffffffffffffffffffffffffffffffffffff;
 
     /// @notice Use when you _really_ really _really_ don't trust the called
     /// contract. This prevents the called contract from causing reversion of
@@ -19,14 +19,12 @@ library ExcessivelySafeCall {
     /// @param _calldata The data to send to the remote contract
     /// @return success and returndata, as `.call()`. Returndata is capped to
     /// `_maxCopy` bytes.
-    function excessivelySafeCall(
-        address _target,
-        uint _gas,
-        uint16 _maxCopy,
-        bytes memory _calldata
-    ) internal returns (bool, bytes memory) {
+    function excessivelySafeCall(address _target, uint256 _gas, uint16 _maxCopy, bytes memory _calldata)
+        internal
+        returns (bool, bytes memory)
+    {
         // set up for assembly call
-        uint _toCopy;
+        uint256 _toCopy;
         bool _success;
         bytes memory _returnData = new bytes(_maxCopy);
         // dispatch message to recipient
@@ -34,20 +32,19 @@ library ExcessivelySafeCall {
         // we call via assembly to avoid memcopying a very large returndata
         // returned by a malicious contract
         assembly {
-            _success := call(
-                _gas, // gas
-                _target, // recipient
-                0, // ether value
-                add(_calldata, 0x20), // inloc
-                mload(_calldata), // inlen
-                0, // outloc
-                0 // outlen
-            )
+            _success :=
+                call(
+                    _gas, // gas
+                    _target, // recipient
+                    0, // ether value
+                    add(_calldata, 0x20), // inloc
+                    mload(_calldata), // inlen
+                    0, // outloc
+                    0 // outlen
+                )
             // limit our copy to 256 bytes
             _toCopy := returndatasize()
-            if gt(_toCopy, _maxCopy) {
-                _toCopy := _maxCopy
-            }
+            if gt(_toCopy, _maxCopy) { _toCopy := _maxCopy }
             // Store the length of the copied bytes
             mstore(_returnData, _toCopy)
             // copy the bytes from returndata[0:_toCopy]
@@ -71,14 +68,13 @@ library ExcessivelySafeCall {
     /// @param _calldata The data to send to the remote contract
     /// @return success and returndata, as `.call()`. Returndata is capped to
     /// `_maxCopy` bytes.
-    function excessivelySafeStaticCall(
-        address _target,
-        uint _gas,
-        uint16 _maxCopy,
-        bytes memory _calldata
-    ) internal view returns (bool, bytes memory) {
+    function excessivelySafeStaticCall(address _target, uint256 _gas, uint16 _maxCopy, bytes memory _calldata)
+        internal
+        view
+        returns (bool, bytes memory)
+    {
         // set up for assembly call
-        uint _toCopy;
+        uint256 _toCopy;
         bool _success;
         bytes memory _returnData = new bytes(_maxCopy);
         // dispatch message to recipient
@@ -86,19 +82,18 @@ library ExcessivelySafeCall {
         // we call via assembly to avoid memcopying a very large returndata
         // returned by a malicious contract
         assembly {
-            _success := staticcall(
-                _gas, // gas
-                _target, // recipient
-                add(_calldata, 0x20), // inloc
-                mload(_calldata), // inlen
-                0, // outloc
-                0 // outlen
-            )
+            _success :=
+                staticcall(
+                    _gas, // gas
+                    _target, // recipient
+                    add(_calldata, 0x20), // inloc
+                    mload(_calldata), // inlen
+                    0, // outloc
+                    0 // outlen
+                )
             // limit our copy to 256 bytes
             _toCopy := returndatasize()
-            if gt(_toCopy, _maxCopy) {
-                _toCopy := _maxCopy
-            }
+            if gt(_toCopy, _maxCopy) { _toCopy := _maxCopy }
             // Store the length of the copied bytes
             mstore(_returnData, _toCopy)
             // copy the bytes from returndata[0:_toCopy]
@@ -118,7 +113,7 @@ library ExcessivelySafeCall {
      */
     function swapSelector(bytes4 _newSelector, bytes memory _buf) internal pure {
         require(_buf.length >= 4);
-        uint _mask = LOW_28_MASK;
+        uint256 _mask = LOW_28_MASK;
         assembly {
             // load the first word of
             let _word := mload(add(_buf, 0x20))
