+++ src/lzApp/libs/LzLib.sol
@@ -14,11 +14,15 @@ library LzLib {
     // Address type handling
 
     struct AirdropParams {
-        uint airdropAmount;
+        uint256 airdropAmount;
         bytes32 airdropAddress;
     }
 
-    function buildAdapterParams(LzLib.AirdropParams memory _airdropParams, uint _uaGasLimit) internal pure returns (bytes memory adapterParams) {
+    function buildAdapterParams(LzLib.AirdropParams memory _airdropParams, uint256 _uaGasLimit)
+        internal
+        pure
+        returns (bytes memory adapterParams)
+    {
         if (_airdropParams.airdropAmount == 0 && _airdropParams.airdropAddress == bytes32(0x0)) {
             adapterParams = buildDefaultAdapterParams(_uaGasLimit);
         } else {
@@ -27,14 +31,18 @@ library LzLib {
     }
 
     // Build Adapter Params
-    function buildDefaultAdapterParams(uint _uaGas) internal pure returns (bytes memory) {
+    function buildDefaultAdapterParams(uint256 _uaGas) internal pure returns (bytes memory) {
         // txType 1
         // bytes  [2       32      ]
         // fields [txType  extraGas]
         return abi.encodePacked(uint16(1), _uaGas);
     }
 
-    function buildAirdropAdapterParams(uint _uaGas, AirdropParams memory _params) internal pure returns (bytes memory) {
+    function buildAirdropAdapterParams(uint256 _uaGas, AirdropParams memory _params)
+        internal
+        pure
+        returns (bytes memory)
+    {
         require(_params.airdropAmount > 0, "Airdrop amount must be greater than 0");
         require(_params.airdropAddress != bytes32(0x0), "Airdrop address must be set");
 
@@ -44,7 +52,7 @@ library LzLib {
         return abi.encodePacked(uint16(2), _uaGas, _params.airdropAmount, _params.airdropAddress);
     }
 
-    function getGasLimit(bytes memory _adapterParams) internal pure returns (uint gasLimit) {
+    function getGasLimit(bytes memory _adapterParams) internal pure returns (uint256 gasLimit) {
         require(_adapterParams.length == 34 || _adapterParams.length > 66, "Invalid adapterParams");
         assembly {
             gasLimit := mload(add(_adapterParams, 34))
@@ -55,12 +63,7 @@ library LzLib {
     function decodeAdapterParams(bytes memory _adapterParams)
         internal
         pure
-        returns (
-            uint16 txType,
-            uint uaGas,
-            uint airdropAmount,
-            address payable airdropAddress
-        )
+        returns (uint16 txType, uint256 uaGas, uint256 airdropAmount, address payable airdropAddress)
     {
         require(_adapterParams.length == 34 || _adapterParams.length > 66, "Invalid adapterParams");
         assembly {
@@ -81,10 +84,10 @@ library LzLib {
     //---------------------------------------------------------------------------
     // Address type handling
     function bytes32ToAddress(bytes32 _bytes32Address) internal pure returns (address _address) {
-        return address(uint160(uint(_bytes32Address)));
+        return address(uint160(uint256(_bytes32Address)));
     }
 
     function addressToBytes32(address _address) internal pure returns (bytes32 _bytes32Address) {
-        return bytes32(uint(uint160(_address)));
+        return bytes32(uint256(uint160(_address)));
     }
 }
